+++ .gitignore
@@ -30,4 +30,7 @@ e2e-results/
 .idea
 
 provisioning/
+cypress/screenshots/actual
+cypress/videos
+cypress/report.json
 __debug_bin
