+++ cypress/integration/utils.ts
@@ -1,0 +1,2 @@
+export const PLUGIN_NAME = 'AstraDB';
+export const PROVISIONING_FILE_NAME = 'datasources/astradb.yaml';
