+++ package.json
@@ -8,6 +8,9 @@
     "test": "grafana-toolkit plugin:test",
     "dev": "grafana-toolkit plugin:dev",
     "watch": "grafana-toolkit plugin:dev --watch",
+    "e2e": "grafana-e2e run",
+    "e2e:open": "grafana-e2e open",
+    "e2e:update": "grafana-e2e run --update-screenshots",
     "sign": "grafana-toolkit plugin:sign",
     "start": "yarn watch",
     "prettier:write": "prettier --list-different \"**/*.{js,ts,tsx,scss,md,mdx}\" --write"
@@ -23,9 +26,12 @@
     "sql-formatter-plus": "^1.3.6"
   },
   "devDependencies": {
+    "@grafana/e2e": "8.4.7",
+    "@grafana/e2e-selectors": "8.4.7",
     "@grafana/toolkit": "8.4.7",
     "@types/lodash": "^4.14.184",
-    "cspell": "^6.8.1"
+    "cspell": "^6.8.1",
+    "cypress": "7.7.0"
   },
   "resolutions": {
     "rxjs": "7.3.0"
